# internal/apidocs/openapi.yaml
openapi: "3.1.0"
info:
  title: VTA Farm API
  version: "1.0.0"
  description: |
    REST API for managing VTA setup sessions with per-user namespace isolation.

    ## Authentication

    All accounts use **passkey-only** authentication. There are no passwords.
    Protected endpoints use **httpOnly cookies** set at login.
    Cookie `vtafarm_user` is required for User endpoints; `vtafarm_admin` for Admin endpoints.

    ### In Scalar / browser
    1. Call the passkey login begin endpoint for your account type
    2. Complete the passkey challenge
    3. The browser stores the cookie automatically — all subsequent requests carry it

    ### curl / non-browser clients
    Use the `token` field from the login response body as a Bearer token:
    ```
    Authorization: Bearer <token>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vtafarm-api.firstperson.dev
    description: Production
...
components:
  securitySchemes:
    CookieAuthUser:
      type: apiKey
      in: cookie
      name: vtafarm_user
    CookieAuthAdmin:
      type: apiKey
      in: cookie
      name: vtafarm_admin
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CreateSetupRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [vta_only, full_stack]
          default: vta_only
        vta_name:
          type: string
          default: "personal-vta"
          example: "personal-vta"
          description: Unique name per user. Auto-generates the DID hosting path {unique_id}/{vta_name}.
        vta_image:
          type: string
          description: Full image URL selected from GET /setup/images.
          example: "ghcr.io/ic3software/vta:0.5.0"
        portable:
          type: boolean
          default: true
          description: Advanced — VTA DID portability. Defaults to true if omitted.
        pre_rotation_count:
          type: integer
          default: 1
          description: Advanced — number of pre-rotation keys. Defaults to 1 if omitted.
      required: [mode, vta_image]
    SetupSessionItem:
      type: object
      properties:
        id:
          type: string
          example: "a1b2c3d4"
        status:
          type: string
          enum: [dns_provisioned, vta_setup_running, vta_setup_complete, provisioning, running, failed]
        mode:
          type: string
          enum: [vta_only, full_stack]
        url:
          type: string
          example: "https://cp-abc12345.example.com"
        vta_name:
          type: string
        vta_did:
          type: string
          description: Populated after vta setup completes
        mediator_did:
          type: string
          description: Populated after vta setup completes
        error_msg:
          type: string
          description: Present only when status is `failed`
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GetSetupSessionResponse:
      type: object
      properties:
        id:
          type: string
          example: "a1b2c3d4"
        status:
          type: string
          enum:
            - dns_provisioned
            - vta_setup_running
            - vta_setup_complete
            - provisioning
            - running
            - failed
        mode:
          type: string
          enum: [vta_only, full_stack]
        url:
          type: string
          example: "https://cp-abc12345.example.com"
        error_msg:
          type: string
          description: Present only when status is `failed`
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSetupResponse:
      type: object
      properties:
        id:
          type: string
          example: "a1b2c3d4"
        url:
          type: string
          example: "https://abc123.example.com"
        status:
          type: string
          example: "dns_provisioned"
paths:
...
  /api/v1/setup:
    post:
      summary: Create setup session
      description: Generates a random subdomain, creates a Cloudflare DNS A record, stores the session, and begins async VTA setup.
      tags: [User]
      security:
        - CookieAuthUser: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSetupRequest"
      responses:
        "201":
          description: Setup session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSetupResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Cluster ingress IP not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Failed to create DNS record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Cloudflare not configured on server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List setup sessions
      description: Returns all setup sessions owned by the requesting user, newest first.
      tags: [User]
      security:
        - CookieAuthUser: []
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetupSessionItem"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/setup/{id}:
    get:
      summary: Get setup session
      description: Returns the current status of a setup session. Poll this endpoint to track progress.
      tags: [User]
      security:
        - CookieAuthUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSetupSessionResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete setup session
      description: Removes the Cloudflare DNS record and deletes the session.
      tags: [User]
      security:
        - CookieAuthUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Session deleted
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Failed to delete DNS record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...